🎨 코드 구조변경 AccountPage & SignUpPage 컴포넌트 최소화
