+++ src/components/AccountPage/getApi.tsx
@@ -1,0 +1,4 @@
+import axios from "axios";
+
+export const axiosData = axios.get("./src/test/graphTest.json");
+export const useTest = axios.get("./src/test/useTest.json");
